Convert to pytest
