+++ .github/workflows/h2o_benchmark.yml
@@ -49,4 +49,4 @@ jobs:
           DASK_COILED__TOKEN: ${{ secrets.COILED_BENCHMARK_BOT_TOKEN }}
           SOFTWARE_ENV: dask-engineering/${{ matrix.software-env }}
         shell: bash -l {0}
-        run: python h2o_benchmark/groupby_h2o.py
+        run: pytest tests/test_h2o_benchmarks.py
