+++ h2o_benchmark/envs/coiled_dist-py38.yml
@@ -4,3 +4,4 @@ channels:
 dependencies:
     - python=3.8
     - coiled-runtime
+    - pytest
