+++ h2o_benchmark/envs/coiled_dist-py39.yml
@@ -4,3 +4,4 @@ channels:
 dependencies:
     - python=3.9
     - coiled-runtime
+    - pytest
