+++ setup.cfg
@@ -3,3 +3,5 @@ max-line-length = 120
 
 [tool:pytest]
 addopts = -v -rsxfE --durations=10 --color=yes
+log_cli = true
+log_level = INFO
